Add files via upload
This file contains Dutch cities, municipalities and provinces and is based on the .csv from https://github.com/bobdenotter/4pp. Since I was not interested in zipcodes but wanted to use it to do a guess about the nearest city, I've removed the zipcode columns. I have updated the data with all the administrative changes in the Netherlands until 11 oct 2023. Or at least I hope I got them all. Use it as you like.